Remove trailing coma from municipality and other possible fields

+++ src/scrapping/speciesLink.py
@@ -5,6 +5,7 @@ from bs4.dammit import EncodingDetector
 
 import requests
 import pandas as pd
+import os
 
 urlSearchTemplate = "http://www.splink.org.br/mod_perl/searchHint?ts_genus={}&offset={}"
 
@@ -71,28 +72,23 @@ def getData(searchTerm, offset=0, inputFile='../data/ListaMacrofitasResult.csv')
 
 def parseDiv(div):
     try:
-            generic_name = div.find("span", {"class":"tGa"}).text.encode('utf8')
+            generic_name = div.find("span", {"class":"tGa"}).text.replace(',', '').encode('utf8')
     except:
             generic_name = ''
     try:
-            specie_name = div.find("span", {"class":"tEa"}).text.encode('utf8').strip()
+            specie_name = div.find("span", {"class":"tEa"}).text.replace(',', '').encode('utf8').strip()
     except:
             specie_name = ''
     try:
-            credit = div.find("span", {"id":"credit"}).text.replace('&#169; ', '').encode('utf8').strip()
-    except:
-            credit = ''
-
-    try:
-            municipality = div.find("span", {"class":"lM"}).text.encode('utf8').strip()
+            municipality = div.find("span", {"class":"lM"}).text.replace(',', '').encode('utf8').strip()
     except:
             municipality = ''
     try:
-            state = div.find("span", {"class":"lS"}).text.encode('utf8').strip()
+            state = div.find("span", {"class":"lS"}).text.replace(',', '').encode('utf8').strip()
     except:
             state = ''
     try:
-            country = div.find("span", {"class":"lC"}).text.encode('utf8').strip()
+            country = div.find("span", {"class":"lC"}).text.replace(',', '').encode('utf8').strip()
     except:
             country = ''
     try:
@@ -112,11 +108,11 @@ def parseDiv(div):
     return scientificName, municipality, state, country, latitude, longitude, date
 
 def writeOutput(registries, outputPath='../data/speciesLink.csv', notFoundPath='../data/notFoundSPLK.csv'):
-    try:
+    if os.path.isfile(outputPath):
         outputLocation = open(outputPath, 'a')
-    except:
+    else:
         outputLocation = open(outputPath, 'w')
-        outputLocation.write("scientificName,municipality,state,country,latitude,longitude,date")
+        outputLocation.write("scientificName,municipality,state,country,latitude,longitude,date\n")
 
     for i in registries:
         outputLocation.write(i + "\n")
